Move common test structs to shared lib test module
