+++ src/lib.rs
@@ -103,3 +103,26 @@ pub fn select_coins<Utxo: WeightedUtxo>(
         select_coins_srd(target, fee_rate, weighted_utxos, &mut thread_rng())
     }
 }
+
+#[cfg(test)]
+mod tests {
+    use bitcoin::{ScriptBuf, TxOut};
+
+    use super::*;
+
+    #[derive(Debug)]
+    pub struct Utxo {
+        pub output: TxOut,
+        pub satisfaction_weight: Weight,
+    }
+
+    pub fn build_utxo(amt: Amount, satisfaction_weight: Weight) -> Utxo {
+        let output = TxOut { value: amt, script_pubkey: ScriptBuf::new() };
+        Utxo { output, satisfaction_weight }
+    }
+
+    impl WeightedUtxo for Utxo {
+        fn satisfaction_weight(&self) -> Weight { self.satisfaction_weight }
+        fn value(&self) -> Amount { self.output.value }
+    }
+}
